requirements.txt
# MindBridge — Python dependencies
# ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
xgboost>=2.0
shap>=0.45
joblib>=1.3

# API
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0

# LLM
openai>=1.0
